잘 되는데 수정하고 등록하면 위치가 바뀜

+++ Final-project/src/main/java/com/kh/finalproject/experience/model/service/ExperienceServiceImpl.java
@@ -69,13 +69,19 @@ public class ExperienceServiceImpl implements ExperienceService {
 	@Transactional
 	public int updateExperience(Experience exp, ArrayList<Files> fileList, String[] oldFiles) {
 		int result = experienceDao.updateExperience(sqlSession, exp);
+		System.out.println(result);
 		System.out.println("게시글 업데이트 하나");
 		for(String oldFile : oldFiles) {
-			result *= experienceDao.deleteFiles(sqlSession, oldFile);
+			System.out.println("비어있니?");
+			System.out.println(oldFile.isEmpty());
+			System.out.println(oldFile.equals(""));
+			if(!oldFile.isEmpty()) result *= experienceDao.deleteFiles(sqlSession, oldFile);
+			System.out.println(result);
 			System.out.println("문제찾기 파일 딜리트");
 		}
 		for(Files file : fileList) {
 			result *= commonDao.insertFiles(sqlSession, file);
+			System.out.println(result);
 			System.out.println("문제찾기 파일 인서트");
 		}
 		return result;
